Refactor unspec test cases & add more test cases
