+++ torchdynamo/variables/tensor.py
@@ -636,11 +636,11 @@ class UnspecializedNumpyVariable(TensorVariable):
     """
 
     def __init__(self, proxy: torch.fx.Proxy, **kwargs):
-        raw_value = kwargs.pop("raw_value", True)
-        volatile_guard_name = kwargs.pop("volatile_guard_name", None)
+        raw_value = kwargs.pop("raw_value", None)
+        volatile_guard = kwargs.pop("volatile_guard", None)
         super(UnspecializedNumpyVariable, self).__init__(proxy, **kwargs)
         self.raw_value = raw_value
-        self.volatile_guard_name = volatile_guard_name
+        self.volatile_guard = volatile_guard
 
     @classmethod
     def from_tensor_variable(cls, tensor_variable, raw_value):
@@ -656,13 +656,13 @@ class UnspecializedPythonVariable(TensorVariable):
     """
 
     def __init__(self, proxy: torch.fx.Proxy, **kwargs):
-        raw_value = kwargs.pop("raw_value", True)
+        raw_value = kwargs.pop("raw_value", None)
         need_unwrap = kwargs.pop("need_unwrap", True)
-        volatile_guard_name = kwargs.pop("volatile_guard_name", None)
+        volatile_guard = kwargs.pop("volatile_guard", None)
         super(UnspecializedPythonVariable, self).__init__(proxy, **kwargs)
         self.raw_value = raw_value
         self.need_unwrap = need_unwrap
-        self.volatile_guard_name = volatile_guard_name
+        self.volatile_guard = volatile_guard
 
     @classmethod
     def from_tensor_variable(cls, tensor_variable, raw_value, need_unwrap=True):
@@ -679,8 +679,7 @@ class UnspecializedPythonVariable(TensorVariable):
                 graph_arg.erase()
 
         for g in self.guards:
-            if g.name == self.volatile_guard_name:
-                assert g.create_fn is GuardBuilder.TYPE_MATCH
+            if g is self.volatile_guard:
                 g.create_fn = GuardBuilder.CONSTANT_MATCH
 
         return ConstantVariable(value=self.raw_value, guards=self.guards)
